Move CNB business logic from file -> platform package
